Split all unit tests to crates/*/src/tests
