+++ api/crates/domain/src/tests.rs
@@ -1,5 +1,5 @@
 mod mocks;
 
-mod external_services;
-mod media;
-mod tags;
+mod entity;
+mod service;
+mod repository;
