+++ api/crates/domain/src/tests/entity.rs
@@ -1,0 +1,5 @@
+mod external_services;
+mod objects;
+mod sources;
+mod tag_types;
+mod tags;
