+++ api/crates/domain/src/tests/service.rs
@@ -1,0 +1,5 @@
+use super::mocks;
+
+mod external_services;
+mod media;
+mod tags;
