+++ api/crates/postgres/src/tests.rs
@@ -1,0 +1,1 @@
+mod sources;
